Survey service implemented

+++ src/main/java/com/example/survey/service/RoleService.java
@@ -1,6 +1,9 @@
 package com.example.survey.service;
 
 import com.example.survey.dto.*;
+import com.example.survey.exception.NoSuchRoleExistsException;
+import com.example.survey.exception.RoleAlreadyExistsException;
+import com.example.survey.exception.RoleIsInUseException;
 import com.example.survey.model.RegisteredUser;
 import com.example.survey.model.Role;
 import com.example.survey.repository.RegisteredUserRepository;
@@ -9,6 +12,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -17,46 +22,78 @@ public class RoleService {
     private final RoleRepository roleRepository;
     private final RegisteredUserRepository registeredUserRepository;
 
+    private final RegisteredUserService registeredUserService;
+
+    //sprawdza czy rola o danej nazwie juz istnieje, jesli tak wyrzuca wyjatek
     public Role addRole(RoleNameDto roleNameDto) {
-        Role role = new Role();
-        role.setName(roleNameDto.name());
-        return roleRepository.save(role);
+        Boolean roleExists = roleRepository.existsByName(roleNameDto.name());
+        if(!roleExists){
+            Role role = new Role();
+            role.setName(roleNameDto.name());
+            return roleRepository.save(role);
+        }
+        else
+            throw new RoleAlreadyExistsException("Role already exists");
     }
 
-    public Role getRoleByName(String name){
-       return roleRepository.findDistinctByName(name);
-    }
+//    public Role getRoleByName(String name){
+//       return roleRepository.findByName(name);
+//    }
 
+    //nie ma wyjatkow, zwroci pusta liste
     public List<Role> getRoles(){
         return roleRepository.findAll();
     }
 
+    //sprawdza czy istnieje, jesli tak to usuwa
     public void deleteRole(Long id){
+        Role role = roleRepository.findById(id).orElseThrow(() -> new NoSuchRoleExistsException(
+                "No role present with id = "+id));
+        if(roleInUse(id))
+            throw new RoleIsInUseException("Role "+role.getName()+" is still in use");
         roleRepository.deleteById(id);
     }
 
+    //najpierw sprawdza czy rola istnieje po id
     public Role updateRole(RoleDto roleDto) {
-        Role role = new Role();
-        role.setId(roleDto.id());
+        Role role = roleRepository.findById(roleDto.id())
+                .orElseThrow(() -> new NoSuchRoleExistsException("No role present with id = "+roleDto.id()));
         role.setName(roleDto.name());
         return roleRepository.save(role);
     }
 
     public RegisteredUser setRegisteredUserRole(RegisteredUserRoleDto registeredUserRoleDto){
         RegisteredUser registeredUser = registeredUserRepository.findDistinctById(registeredUserRoleDto.userId());
-        Role role = roleRepository.findDistinctById(registeredUserRoleDto.roleId());
+
+        //wyjatek tak naprawde obsluzony linijke wyzej, z wlasciwym userem rola nigdy null
+        Role role = roleRepository.findById(registeredUserRoleDto.roleId()).orElseThrow();
         registeredUser.setRole(role);
         return registeredUserRepository.save(registeredUser);
     }
 
+    //sprawdza czy rola z danym id istnieje
+    public Role getRoleById(Long id) {
+        return roleRepository.findById(id)
+                .orElseThrow(() -> new NoSuchRoleExistsException("No role present with id = "+id));
+    }
 
-    //zrobic obsluge bledow
+    private Boolean roleInUse(Long id){
+        Set<RegisteredUser> registeredUsersUsingRole = registeredUserService.getRegisteredUsers()
+                .stream()
+                .filter(registeredUser -> registeredUser.getRole().getId() == id)
+                .collect(Collectors.toSet());
+        return !registeredUsersUsingRole.isEmpty();
+    }
 
-    //setRegisteredUserRole() ?? dodawanie po samym id albo nazwie?
+
+    //todo zrobic obsluge wyjatkow
+
+    //setRegisteredUserRole() +
     //createRole(): addRole() +
     //updateRole() +
     //deleteRole() +
     //getRole(): getSingleRole() +
+    //getRoleById() +
 
     //dodatkowo getRoles() +
 }
